backend/requirements.txt
python-dotenv==1.0.0
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
Pillow==11.0.0
groq==0.13.0
requests==2.32.3
